Refactoring changes and bug fixes
Improved refactoring and fixed all of the errors that I had immediately after the refactor. Issues remaining are the cog_check on the channels command, which will make it impossible to add a disallowed channel to the allowed channels. There is also an issue with module imports whose solution isn't immediately obvious to me. Aside from that, everything should be good. There are also a great deal of commands that should be restricted to administrators that currently aren't.